+++ tests/test_pyppeteer.py
@@ -1116,8 +1116,8 @@ a + b
 
         async def request_check(req):
             self.assertIn('empty', req.url)
-            self.assertEqual(req.method, 'GET')
             self.assertTrue(req.headers.get('user-agent'))
+            self.assertEqual(req.method, 'GET')
             self.assertEqual(req.resourceType, 'document')
             await req.continue_()