+++ CHANGES.md
@@ -12,6 +12,7 @@ History
 * Implement `Frame.content` and `Frame.setContent` methods
 * `page.tracing.start` accepts custom tracing categories option
 * Add `Browser.process` property
+* Add `Request.frame` property
 
 
 ## Version 0.0.14 (2018-03-14)